VC5Decompressor, Array2DRef: consistently use signed 'int' dimensions, -3%
raw.pixls.us-unique/GoPro/HERO6 Black$ ~/rawspeed/build-Clang-Release/googlebenchmark/googlebenchmark-src/tools/compare.py -a benchmarks ~/rawspeed/build-Clang-Release/src/utilities/rsbench/rsbench{-old,} --benchmark_counters_tabular=true --benchmark_min_time=0.00000001 --benchmark_repetitions=3000 GOPR9172.GPR
RUNNING: /home/lebedevri/rawspeed/build-Clang-Release/src/utilities/rsbench/rsbench-old --benchmark_counters_tabular=true --benchmark_min_time=0.00000001 --benchmark_repetitions=3000 GOPR9172.GPR --benchmark_display_aggregates_only=true --benchmark_out=/tmp/tmpjFMxtv
2018-11-25 22:17:42
Running /home/lebedevri/rawspeed/build-Clang-Release/src/utilities/rsbench/rsbench-old
Run on (8 X 4013.39 MHz CPU s)
CPU Caches:
  L1 Data 16K (x8)
  L1 Instruction 64K (x4)
  L2 Unified 2048K (x4)
  L3 Unified 8192K (x1)
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                        Time           CPU Iterations  CPUTime,s CPUTime/WallTime     Pixels Pixels/CPUTime Pixels/WallTime Raws/CPUTime Raws/WallTime WallTime,s
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
GOPR9172.GPR/threads:8/real_time_mean           48 ms         48 ms       3000   0.383351          7.97483        12M       31.3482M        250.017M      2.61235       20.8347  0.0480873
GOPR9172.GPR/threads:8/real_time_median         48 ms         47 ms       3000   0.380017          7.99365        12M       31.5775M        252.478M      2.63146       21.0398   0.047529
GOPR9172.GPR/threads:8/real_time_stddev          2 ms          2 ms       3000  0.0163211         0.113932          0       1081.65k        9.56203M    0.0901377      0.796836   2.40979m
RUNNING: /home/lebedevri/rawspeed/build-Clang-Release/src/utilities/rsbench/rsbench --benchmark_counters_tabular=true --benchmark_min_time=0.00000001 --benchmark_repetitions=3000 GOPR9172.GPR --benchmark_display_aggregates_only=true --benchmark_out=/tmp/tmp0NVkuI
2018-11-25 22:20:10
Running /home/lebedevri/rawspeed/build-Clang-Release/src/utilities/rsbench/rsbench
Run on (8 X 4008.05 MHz CPU s)
CPU Caches:
  L1 Data 16K (x8)
  L1 Instruction 64K (x4)
  L2 Unified 2048K (x4)
  L3 Unified 8192K (x1)
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                        Time           CPU Iterations  CPUTime,s CPUTime/WallTime     Pixels Pixels/CPUTime Pixels/WallTime Raws/CPUTime Raws/WallTime WallTime,s
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
GOPR9172.GPR/threads:8/real_time_mean           47 ms         47 ms       3000   0.371285          7.97539        12M       32.3466M        257.972M      2.69555       21.4976  0.0465675
GOPR9172.GPR/threads:8/real_time_median         46 ms         46 ms       3000     0.3692           7.9938        12M       32.5028M        259.778M      2.70856       21.6482  0.0461932
GOPR9172.GPR/threads:8/real_time_stddev          2 ms          2 ms       3000  0.0118351         0.128403          0       846.219k        7.46788M    0.0705183      0.622324   1.79777m
Comparing /home/lebedevri/rawspeed/build-Clang-Release/src/utilities/rsbench/rsbench-old to /home/lebedevri/rawspeed/build-Clang-Release/src/utilities/rsbench/rsbench
Benchmark                                                 Time             CPU      Time Old      Time New       CPU Old       CPU New
--------------------------------------------------------------------------------------------------------------------------------------
GOPR9172.GPR/threads:8/real_time_pvalue                 0.0000          0.0000      U Test, Repetitions: 3000 vs 3000
GOPR9172.GPR/threads:8/real_time_mean                  -0.0316         -0.0238            48            47            48            47
GOPR9172.GPR/threads:8/real_time_median                -0.0281         -0.0245            48            46            47            46
GOPR9172.GPR/threads:8/real_time_stddev                -0.2540         -0.1021             2             2             2             2

+++ src/librawspeed/decompressors/VC5Decompressor.h
@@ -117,7 +117,7 @@ class VC5Decompressor final : public AbstractDecompressor {
 
   class Wavelet {
   public:
-    uint16_t width, height;
+    int width, height;
     int16_t prescale;
 
     struct AbstractBand {
@@ -188,7 +188,7 @@ class VC5Decompressor final : public AbstractDecompressor {
 
     Wavelet::ReconstructableBand band{/*clampUint*/ true};
     // the final lowband.
-    uint16_t width, height;
+    int width, height;
   };
 
   static constexpr int numChannels = 4;
